chore: rename database.rs into table.rs
