+++ core/src/databases/error.rs
@@ -1,0 +1,13 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum QueryDatabaseError {
+    #[error("{0}")]
+    GenericError(#[from] anyhow::Error),
+    #[error("Too many result rows")]
+    TooManyResultRows,
+    #[error("Result is too large: {0}")]
+    ResultTooLarge(String),
+    #[error("Query execution error: {0}")]
+    ExecutionError(String),
+}
